refactored 2

+++ ft_push_swap_helper.c
@@ -6,7 +6,7 @@
 /*   By: mnououal <mnououal@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/14 21:47:16 by mnououal          #+#    #+#             */
-/*   Updated: 2026/05/16 18:08:07 by mnououal         ###   ########.fr       */
+/*   Updated: 2026/05/16 20:33:48 by mnououal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -95,7 +95,7 @@ static int	init_stack(t_stack *stack, int size)
 static void	manage_adaptive_mode(t_config *config)
 {
 	if (config->mode == ADAPTIVE || config->is_bench_mode)
-		ft_printf("disorder = %d %", 100 * config->disorder);
+		ft_printf("disorder = %d %", 100 * ((int)config->disorder));
 	if (config->mode == ADAPTIVE)
 	{
 		if (config->disorder < 0.2)
